requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.3.0
streamlit==1.28.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
wikipediaapi==0.6.0
